Implementando o controle de esceções globalizado e ajustes finais

+++ src/main/java/com/devsuperior/desafio3/services/ClienteService.java
@@ -2,14 +2,15 @@ package com.devsuperior.desafio3.services;
 
 import com.devsuperior.desafio3.dto.ClienteDTO;
 import com.devsuperior.desafio3.entities.Cliente;
+import com.devsuperior.desafio3.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import com.devsuperior.desafio3.repositories.ClienteRepository;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ClienteService {
@@ -25,8 +26,9 @@ public class ClienteService {
 
     @Transactional(readOnly = true)
     public ClienteDTO findById(Long id) {
-        Optional<Cliente> result = repository.findById(id);
-        ClienteDTO clienteDTO = new ClienteDTO(result.get());
+        Cliente cliente = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Cliente não encontrado"));
+        ClienteDTO clienteDTO = new ClienteDTO(cliente);
         return clienteDTO;
     }
 
@@ -38,12 +40,19 @@ public class ClienteService {
 
     @Transactional
     public ClienteDTO update(ClienteDTO clienteDTO, Long id) {
-        Cliente cliente = repository.getReferenceById(id);
-        return getClienteDTOSavingCliente(clienteDTO, cliente);
+        try {
+            Cliente cliente = repository.getReferenceById(id);
+            return getClienteDTOSavingCliente(clienteDTO, cliente);
+        } catch(EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Cliente não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente não encontrado");
+        }
         repository.deleteById(id);
     }
